+++ grafana-plugin/jest.setup.ts
@@ -4,6 +4,8 @@
  */
 import '@testing-library/jest-dom';
 
+import 'plugin/dayjs';
+
 // https://stackoverflow.com/a/66055672
 // https://jestjs.io/docs/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom
 Object.defineProperty(window, 'matchMedia', {