+++ grafana-plugin/src/state/rootBaseStore/index.ts
@@ -1,7 +1,6 @@
 import { locationService } from '@grafana/runtime';
 import { contextSrv } from 'grafana/app/core/core';
 import { action, computed, makeObservable, observable, runInAction } from 'mobx';
-import moment from 'moment-timezone';
 import qs from 'query-string';
 import { OnCallAppPluginMeta } from 'types';
 
@@ -28,7 +27,7 @@ import { ScheduleStore } from 'models/schedule/schedule';
 import { SlackStore } from 'models/slack/slack';
 import { SlackChannelStore } from 'models/slack_channel/slack_channel';
 import { TelegramChannelStore } from 'models/telegram_channel/telegram_channel';
-import { Timezone } from 'models/timezone/timezone.types';
+import { TimezoneStore } from 'models/timezone/timezone';
 import { UserStore } from 'models/user/user';
 import { UserGroupStore } from 'models/user_group/user_group';
 import { makeRequest } from 'network';
@@ -50,9 +49,6 @@ import FaroHelper from 'utils/faro';
 export class RootBaseStore {
   @observable
   isBasicDataLoaded = false;
-
-  @observable
-  currentTimezone: Timezone = moment.tz.guess() as Timezone;
 
   @observable
   backendVersion = '';
@@ -82,6 +78,9 @@ export class RootBaseStore {
 
   @observable
   incidentFilters: any;
+
+  @observable
+  pageTitle = '';
 
   @observable
   incidentsPage: any = this.initialQuery.p ? Number(this.initialQuery.p) : 1;
@@ -115,6 +114,7 @@ export class RootBaseStore {
   globalSettingStore = new GlobalSettingStore(this);
   filtersStore = new FiltersStore(this);
   labelsStore = new LabelStore(this);
+  timezoneStore = new TimezoneStore(this);
   msteamsChannelStore: MSTeamsChannelStore = new MSTeamsChannelStore(this);
   loaderStore = LoaderStore;
 
@@ -320,6 +320,11 @@ export class RootBaseStore {
         {}
       );
     });
+  }
+
+  @action.bound
+  setPageTitle(title: string) {
+    this.pageTitle = title;
   }
 
   @action